RegisterPayment step moved to proper place
